+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 ChangeLog
 =========
+1.8.0 (2022-10-29)
+-------------------------
+- Added Foreign Key for bookmarks
+- Reworked Move bookmarks
+- Reworked Delete bookmarks
+
+
 1.7.2 (2022-10-25)
 -------------------------
 - Added trigger for delete bookmark
 - Added empty search field
-- Adden Token exchange
+- Added Token exchange
 - Fixed Keyboard shortcuts